requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-rest-framework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-filter==23.4
drf-yasg==1.21.7
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0
aioredis==2.0.1
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
